refine message
